Change decimal fields on positive integer fields fix templates

+++ toys/models.py
@@ -17,7 +17,7 @@ class Toy(models.Model):
     owner = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
     name = models.CharField(max_length=100)
     description = models.TextField()
-    price = models.DecimalField(max_digits=10, decimal_places=2)
+    price = models.PositiveIntegerField()
     stock = models.PositiveIntegerField()
     manufacturer = models.CharField(max_length=100)
     category = models.ManyToManyField(Category, related_name="toys")
